+++ parser/src/ast/newtype_def.rs
@@ -1,4 +1,5 @@
 use super::{Attribute, Ident, TypeName};
+use crate::error::RecursiveNewtype;
 use crate::grammar::Rule;
 use crate::validate::Validate;
 use crate::Span;
@@ -46,6 +47,8 @@ impl NewtypeDef {
     }
 
     pub(crate) fn validate(&self, validate: &mut Validate) {
+        RecursiveNewtype::validate(self, validate);
+
         self.name.validate(validate);
         self.target_type.validate(validate);
     }