+++ parser/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support parsing newtypes of the form `newtype IDENT = TYPE;`.
+- Add error type `RecursiveNewtype`.
 
 ## [0.12.0] - 2025-01-26